Update README.md
After 3 years of independent work, I'm looking to join a company where I can work with a team. I want to be part of something larger than I could every do on my own. If possible, I want to work on a Digital Twin project for a city or locality. I have a background in geocoding and map system as used for public facing web sites. With Digital Cities, there is an opportunity to bring this into 3D space.